reorg the pattern dir
